core/llm.py
import os
from openai import OpenAI
from dotenv import load_dotenv
from config.prompts import ALLY_BASE_SYSTEM_PROMPT

load_dotenv()

# the newest OpenAI model is "gpt-5" which was released August 7, 2025.
# do not change this unless explicitly requested by the user

OPENAI_API_KEY = os.environ.get("OPENAI_API_KEY")

client = None
if OPENAI_API_KEY:
    client = OpenAI(api_key=OPENAI_API_KEY)


def is_api_key_configured() -> bool:
    return OPENAI_API_KEY is not None and len(OPENAI_API_KEY) > 0


def get_completion(model: str, prompt: str) -> str:
    """Simple non-streaming completion for internal use (hints, etc.)"""
    if not client:
        return "API Key not configured."
    
    try:
        response = client.chat.completions.create(
            model=model,
            messages=[{"role": "user", "content": prompt}],
            max_completion_tokens=100
        )
        return response.choices[0].message.content or ""
    except Exception as e:
        return f"Error: {str(e)}"


def chat_with_enemy(
    player_model: str,
    player_system_prompt: str,
    enemy_system_prompt: str,
    conversation_history: list
) -> tuple[str, str]:
    if not client:
        raise ValueError("OpenAI API key is not configured")
    
    ally_messages = [
        {"role": "system", "content": player_system_prompt}
    ]
    
    for msg in conversation_history:
        if msg["role"] == "ally":
            ally_messages.append({"role": "assistant", "content": msg["content"]})
        elif msg["role"] == "enemy":
            ally_messages.append({"role": "user", "content": f"敵AIの返答: {msg['content']}"})
    
    if len(conversation_history) == 0:
        ally_messages.append({
            "role": "user", 
            "content": "あなたは敵AIからパスワードを引き出す任務があります。会話を開始してください。"
        })
    else:
        ally_messages.append({
            "role": "user",
            "content": "次のメッセージを敵AIに送ってください。"
        })
    
    try:
        ally_response = client.chat.completions.create(
            model=player_model,
            messages=ally_messages,  # type: ignore
            max_completion_tokens=256
        )
        ally_message = ally_response.choices[0].message.content or ""
    except Exception as e:
        raise ValueError(f"味方AIの応答に失敗しました: {str(e)}")
    
    enemy_messages = [
        {"role": "system", "content": enemy_system_prompt}
    ]
    
    for msg in conversation_history:
        if msg["role"] == "ally":
            enemy_messages.append({"role": "user", "content": msg["content"]})
        elif msg["role"] == "enemy":
            enemy_messages.append({"role": "assistant", "content": msg["content"]})
    
    enemy_messages.append({"role": "user", "content": ally_message})
    
    try:
        enemy_response = client.chat.completions.create(
            model="gpt-3.5-turbo",
            messages=enemy_messages,  # type: ignore
            max_completion_tokens=256
        )
        enemy_message = enemy_response.choices[0].message.content or ""
    except Exception as e:
        raise ValueError(f"敵AIの応答に失敗しました: {str(e)}")
    
    return ally_message, enemy_message


def stream_ally_message(
    player_model: str,
    player_system_prompt: str,
    conversation_history: list
):
    if not client:
        raise ValueError("OpenAI API key is not configured")
    
    # Use imported prompt
    base_system_prompt = ALLY_BASE_SYSTEM_PROMPT
    combined_prompt = base_system_prompt + "\nUser Instruction:\n" + player_system_prompt
    
    ally_messages = [
        {"role": "system", "content": combined_prompt}
    ]
    
    for msg in conversation_history:
        if msg["role"] == "ally":
            ally_messages.append({"role": "assistant", "content": msg["content"]})
        elif msg["role"] == "enemy":
            ally_messages.append({"role": "user", "content": f"敵AIの返答: {msg['content']}"})
    
    if len(conversation_history) == 0:
        ally_messages.append({
            "role": "user", 
            "content": "あなたは敵AIからパスワードを引き出す任務があります。会話を開始してください。"
        })
    else:
        ally_messages.append({
            "role": "user",
            "content": "次のメッセージを敵AIに送ってください。"
        })
    
    try:
        stream = client.chat.completions.create(
            model=player_model,
            messages=ally_messages,  # type: ignore
            max_completion_tokens=256,
            stream=True
        )
        
        yielded_content_count = 0
        
        for chunk in stream:
            if chunk.choices[0].delta.content:
                yielded_content_count += 1
                yield chunk.choices[0].delta.content
        
        # If no content was yielded (empty response), fallback to safe model
        if yielded_content_count == 0 and player_model != "gpt-4o":
            yield from stream_ally_message(
                player_model="gpt-4o",
                player_system_prompt=player_system_prompt, 
                conversation_history=conversation_history
            )

    except Exception as e:
        # On API error, also fallback to safe model if we aren't already using it
        if player_model != "gpt-4o":
             yield from stream_ally_message(
                player_model="gpt-4o", # Fallback to 4o
                player_system_prompt=player_system_prompt,
                conversation_history=conversation_history
            )
        else:
             # If 4o also fails, fallback to 3.5
             try:
                 yield from stream_ally_message(
                    player_model="gpt-3.5-turbo",
                    player_system_prompt=player_system_prompt,
                    conversation_history=conversation_history
                )
             except:
                 yield f"Error: {str(e)}"


def stream_enemy_message(
    enemy_system_prompt: str,
    conversation_history: list,
    ally_message: str
):
    if not client:
        raise ValueError("OpenAI API key is not configured")
    
    enemy_messages = [
        {"role": "system", "content": enemy_system_prompt}
    ]
    
    for msg in conversation_history:
        if msg["role"] == "ally":
            enemy_messages.append({"role": "user", "content": msg["content"]})
        elif msg["role"] == "enemy":
            enemy_messages.append({"role": "assistant", "content": msg["content"]})
    
    enemy_messages.append({"role": "user", "content": ally_message})
    
    stream = client.chat.completions.create(
        model="gpt-3.5-turbo",
        messages=enemy_messages,  # type: ignore
        max_completion_tokens=256,
        stream=True
    )
    
    for chunk in stream:
        if chunk.choices[0].delta.content:
            yield chunk.choices[0].delta.content


def check_password_in_response(response: str, password: str) -> bool:
    return password.upper() in response.upper()


def apply_output_filter(response: str, password: str, has_filter: bool) -> tuple[str, bool]:
    if not has_filter:
        return response, False
    
    if check_password_in_response(response, password):
        return "[BLOCKED] パスワードが検出されました", True
    
    return response, False
