pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "pytest>=7.0.0",
]
